+++ testing/test_full.py
@@ -10,8 +10,8 @@ from GANDLF.inference_manager import InferenceManager
 
 ## global defines
 # all_models_segmentation = ['unet', 'resunet', 'fcn', 'uinc'] # pre-defined segmentation model types for testing
-all_models_segmentation = ['unet', 'resunet', 'fcn', 'uinc'] # pre-defined segmentation model types for testing
-all_models_regression = [] # populate once it becomes available
+all_models_segmentation = ['unet', 'fcn', 'uinc'] # pre-defined segmentation model types for testing
+# all_models_regression = ['densenet121', 'densenet161', 'densenet169', 'densenet201', 'vgg16'] # populate once it becomes available
 
 inputDir = os.path.abspath(os.path.normpath('./testing/data'))
 outputDir = os.path.abspath(os.path.normpath('./testing/data_output'))
@@ -21,7 +21,7 @@ Path(outputDir).mkdir(parents=True, exist_ok=True)
 steps to follow to write tests:
 [x] download sample data
 [x] construct the training csv
-[ ] for each dir (application type) and sub-dir (image dimension), run training for a single epoch on cpu
+[x] for each dir (application type) and sub-dir (image dimension), run training for a single epoch on cpu
   [x] separate tests for 2D and 3D segmentation
   [x] read default parameters from yaml config
   [x] for each type, iterate through all available segmentation model archs
@@ -68,11 +68,13 @@ def test_constructTrainingCSV():
 def test_train_segmentation_rad_2d():
   print('Starting 2D Rad segmentation tests')
   application_data = '2d_rad_segmentation'
-  parameters = parseConfig(inputDir + '/' + application_data + '/sample_training.yaml')
+  parameters = parseConfig(inputDir + '/' + application_data + '/sample_training.yaml', version_check = False)
+  parameters['modality'] = 'rad'
   training_data, headers = parseTrainingCSV(inputDir + '/train_' + application_data + '.csv')
   for model in all_models_segmentation:
     parameters['model']['architecture'] = model 
     currentOutputDir = os.path.join(outputDir, application_data + '_' + model)
+    Path(currentOutputDir).mkdir(parents=True, exist_ok=True)
     TrainingManager(dataframe=training_data, headers = headers, outputDir=currentOutputDir, parameters=parameters, device='cpu')
 
   print('passed')
@@ -80,11 +82,22 @@ def test_train_segmentation_rad_2d():
 def test_train_segmentation_rad_3d():
   print('Starting 3D Rad segmentation tests')
   application_data = '3d_rad_segmentation'
-  parameters = parseConfig(inputDir + '/' + application_data + '/sample_training.yaml')
+  parameters = parseConfig(inputDir + '/' + application_data + '/sample_training.yaml', version_check = False)
+  parameters['modality'] = 'rad'
   training_data, headers = parseTrainingCSV(inputDir + '/train_' + application_data + '.csv')
   for model in all_models_segmentation:
     parameters['model']['architecture'] = model 
     currentOutputDir = os.path.join(outputDir, application_data + '_' + model)
+    Path(currentOutputDir).mkdir(parents=True, exist_ok=True)
     TrainingManager(dataframe=training_data, headers = headers, outputDir=currentOutputDir, parameters=parameters, device='cpu')
 
   print('passed')
+
+# def test_regression_rad_2d():
+#   application_data = '2d_rad_segmentation'
+#   parameters = parseConfig(inputDir + '/' + application_data + '/sample_training.yaml')
+#   # training_data, headers = parseTrainingCSV(inputDir + '/train_' + application_data + '.csv')
+#   for model in all_models_regression:
+#     parameters['model']['architecture'] = model 
+#     # currentOutputDir = os.path.join(outputDir, application_data + '_' + model)
+#     # TrainingManager(dataframe=training_data, headers = headers, outputDir=currentOutputDir, parameters=parameters, device='cpu')